+++ lib/visual_snapshots/run-web-dextrose.sh
@@ -5,7 +5,6 @@ npx selenium-standalone install &&
 npx selenium-standalone start 2>> selenium.log & 
 npm run fetch-fonts
 
-
 ##construct list of changed packages
 npx lerna ls --json --since > fructose/changedPackages.json
 
@@ -15,6 +14,9 @@ node fructose/changedComponents.js
 ## create components file of all showcase file following pattern
 npx rnscl --pattern '*.showcase*' --outputFile ./fructose/components.js --config ./fructose/rnscl.config
 
+## build vendor package
+npx webpack --config fructose/vendor.webpack.config.js
+
 ##start bundler
 npx react-native start --reset-cache &
 
@@ -23,5 +25,5 @@ npx fructose-web -d ./fructose >> webpack.log &
 FRUCTOSE_WEB_PID=$!
 
 ##run dextrose
-npx dextrose run --config ./dextrose/dextrose.web.js --snapshotWait 2000
+npx dextrose run --config ./dextrose/dextrose.web.js 
 kill -9 $FRUCTOSE_WEB_PID # close the web app