+++ packages/provider/src/topic-articles.js
@@ -2,7 +2,12 @@ import gql from "graphql-tag";
 import connectGraphql from "./connect";
 
 export const query = gql`
-  query TopicQuery($slug: Slug!, $first: Int, $skip: Int, $imageRatio: Ratio!) {
+  query TopicArticlesQuery(
+    $slug: Slug!
+    $first: Int
+    $skip: Int
+    $imageRatio: Ratio!
+  ) {
     topic(slug: $slug) {
       articles(first: $first, skip: $skip) {
         count