+++ packages/topic/__tests__/topic-head-functional.js
@@ -1,14 +1,12 @@
-import "react-native";
 import React from "react";
 import renderer from "react-test-renderer";
 import TopicHead from "../src/topic-head";
 
-jest.mock("../src/topic-head-loading", () => "TopicHeadLoading");
-jest.mock("../src/topic-head-divider", () => "Divider");
-
 export default () => {
   it("should render correctly", () => {
-    const tree = renderer.create(<TopicHead name="Animals" />).toJSON();
+    const tree = renderer
+      .create(<TopicHead isLoading={false} name="Animals" />)
+      .toJSON();
 
     expect(tree).toMatchSnapshot();
   });
@@ -17,6 +15,7 @@ export default () => {
     const tree = renderer
       .create(
         <TopicHead
+          isLoading={false}
           name="Animals"
           description="Animals are multicellular eukaryotic organisms."
         />
@@ -27,9 +26,7 @@ export default () => {
   });
 
   it("should render correctly when loading", () => {
-    const tree = renderer
-      .create(<TopicHead name="Animals" isLoading />)
-      .toJSON();
+    const tree = renderer.create(<TopicHead name="Animals" />).toJSON();
 
     expect(tree).toMatchSnapshot();
   });