+++ packages/topic/src/topic-tracking-context.js
@@ -1,4 +1,3 @@
-import get from "lodash.get";
 import { withTrackingContext } from "@times-components/tracking";
 
 export default Component =>
@@ -6,7 +5,6 @@ export default Component =>
     trackingObjectName: "Topic",
     getAttrs: ({ topic, page, pageSize }) => ({
       topicName: topic && topic.name,
-      articlesCount: get(topic, "articles.count", 0),
       page,
       pageSize
     })