+++ packages/topic/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.2.2"></a>
+## [0.2.2](https://github.com/newsuk/times-components/compare/@times-components/topic@0.2.1...@times-components/topic@0.2.2) (2018-05-23)
+
+
+### Bug Fixes
+
+* Get Topic "Viewed" analytics working ([#954](https://github.com/newsuk/times-components/issues/954)) ([1507be4](https://github.com/newsuk/times-components/commit/1507be4))
+
+
+
+
+<a name="0.2.1"></a>
+## [0.2.1](https://github.com/newsuk/times-components/compare/@times-components/topic@0.2.0...@times-components/topic@0.2.1) (2018-05-23)
+
+
+
+
+**Note:** Version bump only for package @times-components/topic
+
+<a name="0.2.0"></a>
+# [0.2.0](https://github.com/newsuk/times-components/compare/@times-components/topic@0.1.4...@times-components/topic@0.2.0) (2018-05-22)
+
+
+### Features
+
+* REPLAT-2167 Topic name query ([#956](https://github.com/newsuk/times-components/issues/956)) ([f398f51](https://github.com/newsuk/times-components/commit/f398f51))
+
+
+
+
+<a name="0.1.4"></a>
+## [0.1.4](https://github.com/newsuk/times-components/compare/@times-components/topic@0.1.3...@times-components/topic@0.1.4) (2018-05-21)
+
+
+
+
+**Note:** Version bump only for package @times-components/topic
+
+<a name="0.1.3"></a>
+## [0.1.3](https://github.com/newsuk/times-components/compare/@times-components/topic@0.1.2...@times-components/topic@0.1.3) (2018-05-21)
+
+
+
+
+**Note:** Version bump only for package @times-components/topic
+
 <a name="0.1.2"></a>
 ## [0.1.2](https://github.com/newsuk/times-components/compare/@times-components/topic@0.1.1...@times-components/topic@0.1.2) (2018-05-21)